Added time measurement

+++ src/Selector.h
@@ -21,7 +21,7 @@ public:
 
 	virtual bool select(std::vector<Genome*>& genomes, 
                             std::vector< Genome* >& selected,
-                            int sum_fitness ) = 0;
+                            unsigned long int sum_fitness ) = 0;
 protected:
     unsigned int m_nr_combine;
     unsigned int m_nr_die;
@@ -32,7 +32,7 @@ public:
 	RouletteSelector( unsigned int c, unsigned int d ) : Selector( c, d ) { };
 	bool select(std::vector<Genome*>& population,
 							std::vector< Genome* >& selected,
-							int sum_fitness );
+							unsigned long int sum_fitness );
 };
 
 class FittestSelector : public Selector {
@@ -40,7 +40,7 @@ public:
 	FittestSelector( unsigned int c, unsigned int d ) : Selector( c, d ) { };
 	bool select(std::vector<Genome*>& population,
 							std::vector< Genome* >& selected,
-							int sum_fitness );
+							unsigned long int sum_fitness );
 };
 
 class TournamentSelector : public Selector {
@@ -48,7 +48,7 @@ public:
 	TournamentSelector( unsigned int c, unsigned int d ) : Selector( c, d ) { };
 	bool select(std::vector<Genome*>& population,
 							std::vector< Genome* >& selected,
-							int sum_fitness );
+							unsigned long int sum_fitness );
 };
 
 #endif /* SELECTOR_H_ */
